CARP-710 improvement: Deletion of a group is now a soft delete

+++ src/API/RestService/RestApi/Commands/RideCommands/DeleteRideCommand.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading;
+﻿using System;
+using System.Threading;
 using System.Threading.Tasks;
 using AutoWrapper.Wrappers;
 using Domain.Contracts;
@@ -44,6 +45,9 @@ namespace RestApi.Commands.RideCommands
 
 			if (ride.OwnerId != request.AppUserId)
 				throw new ApiException("User cannot delete other user ride.", StatusCodes.Status403Forbidden);
+
+			if (ride.Date >= DateTimeOffset.Now)
+				throw new ApiException("Cannot remove past ride.", StatusCodes.Status403Forbidden);
 			
 			_rideRepository.Delete(ride);
